Creating bitcoinj test util
