+++ src/test/scala/org/scalacoin/crypto/TransactionSignatureSerializerTest.scala
@@ -47,14 +47,14 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
   it must "serialize a transaction for SIGHASH_ALL correctly" in {
     require(scriptPubKey.hex == BitcoinSUtil.encodeHex(multiSigScript.getProgram), "Script pub key hex not the same as multiSigScript hex")
 
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     println(BitcoinSUtil.encodeHex(multiSigScript.getProgram))
     val sigBytes : Seq[Byte] = TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey,SIGHASH_ALL)
     val bitcoinjSerialization = BitcoinSUtil.encodeHex(
-      BitcoinJSignatureSerialization.serializeForSignature(bitcoinjMultiSigTransaction,0,multiSigScript.getProgram(),SIGHASH_ALL.byte)
+      BitcoinJSignatureSerialization.serializeForSignature(BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram(),SIGHASH_ALL.byte)
     )
 
     BitcoinSUtil.encodeHex(sigBytes) must be (bitcoinjSerialization)
@@ -62,180 +62,170 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
 
   it must "hash a transction with SIGHASH_ALL correctly" in {
 
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val bitcoinsTxSigHash : Seq[Byte] = TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey,SIGHASH_ALL)
     val bitcoinjTxSigHash = BitcoinSUtil.encodeHex(
-      BitcoinJSignatureSerialization.hashForSignature(bitcoinjMultiSigTransaction,0,multiSigScript.getProgram(),SIGHASH_ALL.byte)
+      BitcoinJSignatureSerialization.hashForSignature(BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram(),SIGHASH_ALL.byte)
     )
     BitcoinSUtil.encodeHex(bitcoinsTxSigHash) must be (bitcoinjTxSigHash)
   }
   it must "serialize a transaction for a SIGHASH_SINGLE transaction correctly" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serialiazedTxForSig : Seq[Byte] = TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey,SIGHASH_SINGLE)
 
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.serializeForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE.byte))
     BitcoinSUtil.encodeHex(serialiazedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "hash a transaction for a SIGHASH_SINGLE signature correctly" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serialiazedTxForSig : Seq[Byte] = TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey,SIGHASH_SINGLE)
 
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE.byte))
     BitcoinSUtil.encodeHex(serialiazedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "serialize a transaction for SIGHASH_NONE signature correctly" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serialiazedTxForSig : Seq[Byte] = TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey,SIGHASH_NONE)
 
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.serializeForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE.byte))
     BitcoinSUtil.encodeHex(serialiazedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "hash a transaction for a SIGHASH_NONE signature correctly" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serialiazedTxForSig : Seq[Byte] = TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey,SIGHASH_NONE)
 
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE.byte))
     BitcoinSUtil.encodeHex(serialiazedTxForSig) must be (bitcoinjSigSerialization)
 
   }
 
   it must "serialize a transaction that has the SIGHASH_ANYONECANPAY flag set" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxForSig : Seq[Byte] = TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey, SIGHASH_ANYONECANPAY)
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.serializeForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxForSig) must be (bitcoinjSigSerialization)
 
   }
 
   it must "hash a transaction for SIGHASH_ANYONECANPAY correctly" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxForSig : Seq[Byte] =
       TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey, SIGHASH_ANYONECANPAY)
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "serialize a transaction that uses both SIGHASH_ANYONECANPAY & SIGHASH_ALL" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxForSig : Seq[Byte] =
       TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey, SIGHASH_ALL_ANYONECANPAY)
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.serializeForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ALL_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ALL_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "serialize a transaction that uses both SIGHASH_ANYONECANPAY & SIGHASH_SINGLE" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxForSig : Seq[Byte] =
       TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey, SIGHASH_SINGLE_ANYONECANPAY)
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.serializeForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "serialize a transaction that uses both SIGHASH_ANYONECANPAY & SIGHASH_NONE" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxForSig : Seq[Byte] =
       TransactionSignatureSerializer.serializeForSignature(spendingTx,0,scriptPubKey, SIGHASH_NONE_ANYONECANPAY)
     val bitcoinjSigSerialization = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.serializeForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxForSig) must be (bitcoinjSigSerialization)
   }
 
   it must "hash a transaction that uses both SIGHASH_ANYONECANPAY & SIGHASH_ALL" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxHashForSig : Seq[Byte] =
       TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey, SIGHASH_ALL_ANYONECANPAY)
     val bitcoinjTxHashForSig = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ALL_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_ALL_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxHashForSig) must be (bitcoinjTxHashForSig)
   }
 
   it must "hash a transaction that uses both SIGHASH_ANYONECANPAY & SIGHASH_SINGLE" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxHashForSig : Seq[Byte] =
       TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey, SIGHASH_SINGLE_ANYONECANPAY)
     val bitcoinjTxHashForSig = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_SINGLE_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxHashForSig) must be (bitcoinjTxHashForSig)
   }
 
   it must "hash a transaction that uses both SIGHASH_ANYONECANPAY & SIGHASH_NONE" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
+    val spendingTx = Transaction.factory(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
+    spendingTx.hex must be (BitcoinSUtil.encodeHex(BitcoinJTestUtil.multiSigTransaction.bitcoinSerialize()))
 
     val serializedTxHashForSig : Seq[Byte] =
       TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey, SIGHASH_NONE_ANYONECANPAY)
     val bitcoinjTxHashForSig = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE_ANYONECANPAY.byte))
+      BitcoinJTestUtil.multiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE_ANYONECANPAY.byte))
 
     BitcoinSUtil.encodeHex(serializedTxHashForSig) must be (bitcoinjTxHashForSig)
   }
 
-  it must "sign an input for a transaction" in {
-    val spendingTx = Transaction.factory(bitcoinjMultiSigTransaction.bitcoinSerialize())
 
-    spendingTx.hex must be (BitcoinSUtil.encodeHex(bitcoinjMultiSigTransaction.bitcoinSerialize()))
-
-    val serializedTxHashForSig : Seq[Byte] =
-      TransactionSignatureSerializer.hashForSignature(spendingTx,0,scriptPubKey, SIGHASH_NONE_ANYONECANPAY)
-    val bitcoinjTxHashForSig = BitcoinSUtil.encodeHex(BitcoinJSignatureSerialization.hashForSignature(
-      bitcoinjMultiSigTransaction,0,multiSigScript.getProgram,SIGHASH_NONE_ANYONECANPAY.byte))
-  }
-
-  it must "serialize a simple transaction with one input for signing" in {
+  it must "serialize a simple transaction with one p2pkh input for signing" in {
     val (spendingTx,spendingInput, inputIndex, creditingOutput) =
       TransactionTestUtil.transactionWithSpendingInputAndCreditingOutput
 
@@ -262,9 +252,6 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
     val serializedTxForSig : String = BitcoinSUtil.encodeHex(
       TransactionSignatureSerializer.serializeForSignature(spendingTx,inputIndex,creditingOutput.scriptPubKey,hashType
       ))
-
-
-
     serializedTxForSig must be (BitcoinSUtil.encodeHex(bitcoinjSerializeForSig))
   }
 
@@ -283,10 +270,6 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
     val parentTx = new org.bitcoinj.core.Transaction(params,Utils.HEX.decode(rawParentTx))
     val parentOutput = parentTx.getOutput(input.getOutpoint.getIndex)
 
-    //connect the input to the output
-    input.connect(parentOutput)
-    val pubKey : Array[Byte]  = scriptSig.getPubKey
-    val signature  : Array[Byte] = scriptSig.getChunks().get(0).data
     val bitcoinjSerializeForSig : Seq[Byte] =
       BitcoinJSignatureSerialization.serializeForSignature(bitcoinjTx,inputIndex,
         parentOutput.getScriptBytes, SIGHASH_ALL.byte)
@@ -295,8 +278,6 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
     val serializedTxForSig : String = BitcoinSUtil.encodeHex(
       TransactionSignatureSerializer.serializeForSignature(spendingTx,inputIndex,creditingOutput.scriptPubKey,hashType
     ))
-    println("Bitcoin-S: " +  serializedTxForSig)
-    println("Bitcoin-J: " + BitcoinSUtil.encodeHex(bitcoinjSerializeForSig))
 
 
     serializedTxForSig must be (BitcoinSUtil.encodeHex(bitcoinjSerializeForSig))
@@ -311,7 +292,7 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
    * @return
    */
   private def createBitcoinjMultiSigScriptHashForSig(hashType : HashType) : String = {
-    val spendTx = bitcoinjMultiSigTransaction
+    val spendTx = BitcoinJTestUtil.multiSigTransaction
 
     val sighash : Seq[Byte] = hashType match {
       case SIGHASH_ALL => spendTx.hashForSignature(0, multiSigScript, SigHash.ALL, false).getBytes
@@ -326,20 +307,6 @@ class TransactionSignatureSerializerTest extends FlatSpec with MustMatchers {
   }
 
 
-  private def bitcoinjMultiSigTransaction : org.bitcoinj.core.Transaction = {
-    //https://github.com/bitcoinj/bitcoinj/blob/master/core/src/test/java/org/bitcoinj/script/ScriptTest.java#L127
-    val txHex = "01000000013df681ff83b43b6585fa32dd0e12b0b502e6481e04ee52ff0fdaf55a16a4ef61000000006b483045022100a84acca7906c13c5895a1314c165d33621cdcf8696145080895cbf301119b7cf0220730ff511106aa0e0a8570ff00ee57d7a6f24e30f592a10cae1deffac9e13b990012102b8d567bcd6328fd48a429f9cf4b315b859a58fd28c5088ef3cb1d98125fc4e8dffffffff02364f1c00000000001976a91439a02793b418de8ec748dd75382656453dc99bcb88ac40420f000000000017a9145780b80be32e117f675d6e0ada13ba799bf248e98700000000"
-    val params = TestNet3Params.get()
-    val creditingTx = new org.bitcoinj.core.Transaction(params,BitcoinSUtil.decodeHex(txHex).toArray)
-    val output = creditingTx.getOutput(1)
-    val spendTx = new org.bitcoinj.core.Transaction(params)
-    val address = new org.bitcoinj.core.Address(params, "n3CFiCmBXVt5d3HXKQ15EFZyhPz4yj5F3H")
-    val outputScript = org.bitcoinj.script.ScriptBuilder.createOutputScript(address)
-    spendTx.addOutput(output.getValue(), outputScript)
-    spendTx.addInput(output)
-    spendTx
-
-  }
